@@ -15,10 +15,10 @@ void VoronoiGrid::OnRender(Scene &scene)
 {
 	TraverseGrid::OnRender(scene);
 
-	_drawFlags &DrawFlag_Grid ? _grid.ShowGrid() : _grid.HideGrid();
-	_drawFlags &DrawFlag_Objects ? _grid.ShowFilled() : _grid.HideFilled();
+	_drawFlags &TraverseGridDrawFlag_Grid ? _grid.ShowGrid() : _grid.HideGrid();
+	_drawFlags &TraverseGridDrawFlag_Objects ? _grid.ShowFilled() : _grid.HideFilled();
 	scene.Submit(_grid);
-	if ( _drawFlags & DrawFlag_Objects )
+	if ( _drawFlags & TraverseGridDrawFlag_Objects )
 	{
 		scene.Submit(_filledEdgesVA);
 	}
@@ -33,19 +33,19 @@ void VoronoiGrid::OnRenderTargetResize(const sf::Vector2f &size)
 
 void VoronoiGrid::ClearNodeColor(int uid)
 {
-	SE_CORE_ASSERT(uid != -1, "Invalid uid");
-	_grid.GetPolygon(GetNode(uid).GetPosition()).SetFillColor(sf::Color::Transparent);
+	Debug::Assert(uid != -1, "Invalid uid");
+	_grid.PolygonAt(NodeByUid(uid).Position()).SetFillColor(sf::Color::Transparent);
 }
 
 void VoronoiGrid::SetNodeColor(int uid, const sf::Color &color)
 {
-	SE_CORE_ASSERT(uid != -1, "Invalid uid");
-	_grid.GetPolygon(GetNode(uid).GetPosition()).SetFillColor(color);
+	Debug::Assert(uid != -1, "Invalid uid");
+	_grid.PolygonAt(NodeByUid(uid).Position()).SetFillColor(color);
 }
 
 void VoronoiGrid::ClearNodeEdgeColor(int fromUid, int toUid)
 {
-	SE_CORE_ASSERT(fromUid != -1 && toUid != -1, "Invalid uid");
+	Debug::Assert(fromUid != -1 && toUid != -1, "Invalid uid");
 	const auto findResult = _filledEdges.find({ fromUid, toUid });
 	if ( findResult != _filledEdges.end() )
 	{
@@ -59,19 +59,19 @@ void VoronoiGrid::ClearNodeEdgeColor(int fromUid, int toUid)
 
 void VoronoiGrid::SetNodeEdgeColor(int fromUid, int toUid, const sf::Color &color)
 {
-	SE_CORE_ASSERT(fromUid != -1 && toUid != -1, "Invalid uid");
+	Debug::Assert(fromUid != -1 && toUid != -1, "Invalid uid");
 	if ( _filledEdges.find({ fromUid, toUid }) == _filledEdges.end() )
 	{
-		const auto &firstPosition = GetNode(fromUid).GetPosition();
-		const auto &secondPosition = GetNode(toUid).GetPosition();
+		const auto &firstPosition = NodeByUid(fromUid).Position();
+		const auto &secondPosition = NodeByUid(toUid).Position();
 
-		const auto firstPolygon = _grid.GetPolygon(firstPosition);
-		const auto secondPolygon = _grid.GetPolygon(secondPosition);
+		const auto firstPolygon = _grid.PolygonAt(firstPosition);
+		const auto secondPolygon = _grid.PolygonAt(secondPosition);
 
-		const auto firstPolygonMid = GenUtils::Mid(firstPolygon.GetPoints());
-		const auto secondPolygonMid = GenUtils::Mid(secondPolygon.GetPoints());
+		const auto firstPolygonMid = GenUtils::Mid(firstPolygon.Points());
+		const auto secondPolygonMid = GenUtils::Mid(secondPolygon.Points());
 
-		const auto closestEdge = firstPolygon.GetClosestEdge(firstPolygonMid + (secondPolygonMid - firstPolygonMid) / 2.0f);
+		const auto closestEdge = firstPolygon.ClosestEdge(firstPolygonMid + (secondPolygonMid - firstPolygonMid) / 2.0f);
 		const auto leftToRightNorm = VecUtils::Unit(closestEdge.first - closestEdge.second);
 
 		const auto perpendicularLeftToRightNorm = VecUtils::Perpendicular(leftToRightNorm);
@@ -108,25 +108,25 @@ void VoronoiGrid::GenerateNodes()
 {
 	_nodes.clear();
 	int uid = 0;
-	const auto &polygons = _grid.GetPolygons();
+	const auto &polygons = _grid.Polygons();
 	for ( const auto &polygon : polygons )
 	{
-		_nodes.emplace(uid++, Node(uid, GenUtils::Mid(polygon.GetPoints())));
+		_nodes.emplace(uid++, Node(uid, GenUtils::Mid(polygon.Points())));
 	}
 	_grid.SetOutlineColor(_gridColor);
 }
 
 void VoronoiGrid::CalculateNeighbors()
 {
-	const auto &polygons = _grid.GetPolygons();
+	const auto &polygons = _grid.Polygons();
 	for ( size_t i = 0; i < polygons.size(); i++ )
 	{
-		const auto &neighbors = polygons[i].GetNeighbors();
+		const auto &neighbors = polygons[i].Neighbors();
 		for ( const auto &neighbor : neighbors )
 		{
-			const int neighborUID = GetNodeUID(GenUtils::Mid(neighbor->GetPoints()));
-			const float cost = VecUtils::Length(GenUtils::Mid(polygons[i].GetPoints()) - GetNode(neighborUID).GetPosition());
-			GetNode(i).AddNeighbor(neighborUID, cost);
+			const int neighborUID = NodeUidByPosition(GenUtils::Mid(neighbor->Points()));
+			const float cost = VecUtils::Length(GenUtils::Mid(polygons[i].Points()) - NodeByUid(neighborUID).Position());
+			NodeByUid(i).AddNeighbor(neighborUID, cost);
 		}
 	}
 }
