@@ -2,30 +2,22 @@
 
 namespace Se
 {
-
-SignalAggregate<const sf::Vector2f &> BaseLayer::Signals::OnRenderTargetResize;
-
 BaseLayer::BaseLayer() :
 	_controllableRenderTexture(100, 100),
 	_scene("Scene", &_controllableRenderTexture, &_camera)
 {
 }
 
-void BaseLayer::OnAttach(std::shared_ptr<BatchLoader> &loader)
+void BaseLayer::OnAttach(Shared<BatchLoader>& loader)
 {
-	_scene.GetViewportPane().GetSignal(ViewportPane::Signals::OnWantRenderTargetResize).Connect(
-		[this](const sf::Vector2f &size)
-		{
-			OnWantRenderTargetResize(size);
-		}
-	);
+	_scene.ViewportPane().Resized += SE_EV_ACTION(BaseLayer::OnWantRenderTargetResize);
 	RenderTargetManager::Add(&_controllableRenderTexture);
 
-	GetSignal(Signals::OnRenderTargetResize).Connect(
-		[this](const sf::Vector2f &size)
-		{
-			OnRenderTargetResize(size);
-		});
+	RenderTargetResized += [this](const sf::Vector2f& newSize)
+	{
+		OnRenderTargetResize(newSize);
+		return false;
+	};
 }
 
 void BaseLayer::OnDetach()
@@ -44,17 +36,16 @@ void BaseLayer::OnPostFrame()
 
 void BaseLayer::OnUpdate()
 {
-	if ( _wantResize )
+	if (_wantResize)
 	{
-		if ( _noResizeRequestTimer.asSeconds() > 0.2f && _framesWithNoResizeRequest > 4 )
+		if (_framesWithNoResizeRequest > 4)
 		{
-			GetSignals().Emit(Signals::OnRenderTargetResize, _resizeTo);
+			RenderTargetResized.Invoke(_resizeTo);
 			_wantResize = false;
 		}
 		else
 		{
 			_framesWithNoResizeRequest++;
-			_noResizeRequestTimer += Global::Clock::GetFrameTime();
 		}
 	}
 
@@ -63,30 +54,27 @@ void BaseLayer::OnUpdate()
 
 void BaseLayer::OnGuiRender()
 {
-	if ( _viewSystem )
-	{
-		_camera.OnGuiRender();
-		_terminal.OnGuiRender();
-		Application::Get().OnGuiRender();
-	}
+	Gui::Instance().OnGuiRender();
+	_camera.OnGuiRender();
+	_terminal.OnGuiRender();
+	App::Instance().OnGuiRender();
 	_scene.OnGuiRender();
 }
 
-void BaseLayer::OnRenderTargetResize(const sf::Vector2f &newSize)
+void BaseLayer::OnRenderTargetResize(const sf::Vector2f& newSize)
 {
-	_controllableRenderTexture.GetRenderTexture().create(newSize.x, newSize.y);
+	_controllableRenderTexture.RenderTexture().create(newSize.x, newSize.y);
 	_camera.SetViewportSize(newSize);
 }
 
-void BaseLayer::OnWantRenderTargetResize(const sf::Vector2f &newSize)
+void BaseLayer::OnWantRenderTargetResize(const sf::Vector2f& newSize)
 {
-	if ( newSize == _resizeTo )
+	if (newSize == _resizeTo)
 	{
 		return;
 	}
 	_wantResize = true;
 	_resizeTo = newSize;
-	_noResizeRequestTimer = sf::Time::Zero;
 	_framesWithNoResizeRequest = 0;
 }
 }
