@@ -2,19 +2,22 @@
 
 namespace Se
 {
-void ProjectLayer::OnAttach(std::shared_ptr<BatchLoader> &loader)
+void ProjectLayer::OnAttach(Shared<BatchLoader> &loader)
 {
 	BaseLayer::OnAttach(loader);
 
 	loader->Submit([this]
 				   {
-					   _pathfinderManager = std::make_shared<PathfinderManager>();
+					   _pathfinderManager = CreateShared<PathfinderManager>();
+					   _pathfinderManager->OnUpdate(_scene);
+					   _pathfinderManager->OnRender(_scene);
 				   }, "Creating Pathfinder Manager");
 
 }
 
 void ProjectLayer::OnDetach()
 {
+	_pathfinderManager->Reset();
 	BaseLayer::OnDetach();
 }
 
@@ -30,33 +33,18 @@ void ProjectLayer::OnGuiRender()
 {
 	BaseLayer::OnGuiRender();
 
-	if ( ImGui::BeginMenuBar() )
-	{
-		if ( ImGui::BeginMenu("Gui") )
-		{
-			ImGui::MenuItem("View System", nullptr, &_viewSystem);
-			ImGui::MenuItem("View Demo", nullptr, &_viewDemo);
-			ImGui::EndMenu();
-		}
-		ImGui::EndMenuBar();
-	}
-
 	if ( ImGui::Begin("Project") )
 	{
 		_pathfinderManager->OnGuiRender();
 	}
 	ImGui::End();
-
-	if ( _viewDemo )
-	{
-		ImGui::ShowDemoWindow();
-	}
 }
 
 void ProjectLayer::OnRenderTargetResize(const sf::Vector2f &newSize)
 {
 	BaseLayer::OnRenderTargetResize(newSize);
 	_scene.OnRenderTargetResize(newSize);
+	
 	_pathfinderManager->OnRenderTargetResize(newSize);
 }
 }
