@@ -2,55 +2,62 @@
 
 namespace Se
 {
+AStar::AStar() :
+	Pathfinder("A*")
+{
+	SetBodyColor(sf::Color::Cyan);
+}
+
 void AStar::FindPath(int startUID, int goalUID)
 {
 	_checkingQueue.push_front(startUID);
-	GetNode(startUID).SetCost("Tentative", 0.0f);
-	while ( !_checkingQueue.empty() && _state != State::BeingCollected )
+	NodeByUid(startUID).SetCost("Tentative", 0.0f);
+	while (!_checkingQueue.empty() && _state != PathfinderState::BeingCollected)
 	{
 		PauseCheck();
 		_activeNodeUID = _checkingQueue.front();
-		if ( _activeNodeUID == goalUID )
+		if (_activeNodeUID == goalUID)
 		{
 			break;
 		}
 
-		Node &activeNode = GetNode(_activeNodeUID);
+		Node& activeNode = NodeByUid(_activeNodeUID);
 		_checkingQueue.pop_front();
 
-		for ( const auto &neighborUID : activeNode.GetNeighbors() )
-		{
-			if ( _state == State::BeingCollected )
-				break;
+		for (const auto& neighborUID : activeNode.Neighbors())
+		{			
+			if (_state == PathfinderState::BeingCollected) return;
 			PauseCheck();
 			SleepDelay();
 
-			Node &neighbor = GetNode(neighborUID);
-			if ( !_traverseGrid->IsEdgeObstacle(_activeNodeUID, neighborUID) && neighborUID != activeNode.GetViaUID() )
+			Node& neighbor = NodeByUid(neighborUID);
+			if (!_traverseGrid->IsEdgeObstacle(_activeNodeUID, neighborUID) && neighborUID != activeNode.ViaUID())
 			{
-				const float suggestedTentativeCost = activeNode.GetCost("Tentative") + activeNode.GetNeighborCost(neighborUID);
-				if ( suggestedTentativeCost < neighbor.GetCost("Tentative") )
+				const float suggestedTentativeCost = activeNode.Cost("Tentative") + activeNode.NeighborCostByUid(
+					neighborUID);
+				if (suggestedTentativeCost < neighbor.Cost("Tentative"))
 				{
-					if ( std::find(_checkingQueue.begin(), _checkingQueue.end(), neighborUID) == _checkingQueue.end() )
+					if (std::ranges::find(_checkingQueue, neighborUID) == _checkingQueue.end())
 					{
 						_checkingQueue.push_back(neighborUID);
 					}
 
 					neighbor.SetVia(_activeNodeUID);
 					neighbor.SetCost("Tentative", suggestedTentativeCost);
-					if ( !neighbor.HasCost("Heuristic") )
+					if (!neighbor.HasCost("Heuristic"))
 					{
-						neighbor.SetCost("Heuristic", VecUtils::Length(neighbor.GetPosition() - GetNode(goalUID).GetPosition()));
+						neighbor.SetCost("Heuristic",
+						                 VecUtils::Length(neighbor.Position() - NodeByUid(goalUID).Position()));
 					}
-					neighbor.SetCost("Total", suggestedTentativeCost + neighbor.GetCost("Heuristic"));
+					neighbor.SetCost("Total", suggestedTentativeCost + neighbor.Cost("Heuristic"));
 				}
 			}
 			activeNode.AddVisitedNeighbor(neighborUID);
 		}
-		std::sort(_checkingQueue.begin(), _checkingQueue.end(), [this](const int &lhs, const int &rhs)
-				  {
-					  return GetNode(lhs).GetCost("Total") < GetNode(rhs).GetCost("Total");
-				  });
+		std::ranges::sort(_checkingQueue, [this](const int& lhs, const int& rhs)
+		{
+			return NodeByUid(lhs).Cost("Total") < NodeByUid(rhs).Cost("Total");
+		});
 	}
 	_checkingQueue.clear();
 }
