@@ -2,35 +2,40 @@
 
 namespace Se
 {
+BestFirstSearch::BestFirstSearch() :
+	Pathfinder("Best First Search")
+{
+	SetBodyColor(sf::Color::Magenta);
+}
+
 void BestFirstSearch::FindPath(int startUID, int goalUID)
 {
 	_checkingQueue.push_front(startUID);
-	GetNode(startUID).SetCost("Heuristic", VecUtils::Length(GetNode(startUID).GetPosition() - GetNode(goalUID).GetPosition()));
-	while ( !_checkingQueue.empty() && _state != State::BeingCollected )
+	NodeByUid(startUID).SetCost("Heuristic",
+	                            VecUtils::Length(NodeByUid(startUID).Position() - NodeByUid(goalUID).Position()));
+	while (!_checkingQueue.empty() && _state != PathfinderState::BeingCollected)
 	{
 		PauseCheck();
 		_activeNodeUID = _checkingQueue.front();
-		if ( _activeNodeUID == goalUID )
+		if (_activeNodeUID == goalUID)
 		{
 			break;
 		}
 
-		Node &activeNode = GetNode(_activeNodeUID);
+		Node& activeNode = NodeByUid(_activeNodeUID);
 		_checkingQueue.pop_front();
 
-		for ( const auto &neighborUID : activeNode.GetNeighbors() )
+		for (const auto& neighborUID : activeNode.Neighbors())
 		{
-			if ( _state == State::BeingCollected )
-				break;
+			if (_state == PathfinderState::BeingCollected) return;
 			PauseCheck();
 			SleepDelay();
 
-			Node &neighbor = GetNode(neighborUID);
-			if ( !_traverseGrid->IsEdgeObstacle(_activeNodeUID, neighborUID) && neighborUID != activeNode.GetViaUID() )
+			Node& neighbor = NodeByUid(neighborUID);
+			if (!_traverseGrid->IsEdgeObstacle(_activeNodeUID, neighborUID) && neighborUID != activeNode.ViaUID())
 			{
-				neighbor.SetCost("Heuristic",
-								 VecUtils::Length(neighbor.GetPosition() - GetNode(goalUID).GetPosition()));
-				if ( !neighbor.WasVisited() )
+				neighbor.SetCost("Heuristic", VecUtils::Length(neighbor.Position() - NodeByUid(goalUID).Position()));
+				if (!neighbor.Visited())
 				{
 					_checkingQueue.push_front(neighborUID);
 					neighbor.SetVia(_activeNodeUID);
@@ -38,12 +43,10 @@ void BestFirstSearch::FindPath(int startUID, int goalUID)
 			}
 			activeNode.AddVisitedNeighbor(neighborUID);
 		}
-		std::sort(_checkingQueue.begin(),
-				  _checkingQueue.end(),
-				  [this](const auto &lhs, const auto &rhs)
-				  {
-					  return GetNode(lhs).GetCost("Heuristic") < GetNode(rhs).GetCost("Heuristic");
-				  });
+		std::ranges::sort(_checkingQueue, [this](const auto& lhs, const auto& rhs)
+		{
+			return NodeByUid(lhs).Cost("Heuristic") < NodeByUid(rhs).Cost("Heuristic");
+		});
 	}
 	_checkingQueue.clear();
 }
