@@ -2,38 +2,43 @@
 
 namespace Se
 {
+Dijkstra::Dijkstra() :
+	Pathfinder("Dijkstra")
+{
+	SetBodyColor(sf::Color::Yellow);
+}
+
 void Dijkstra::FindPath(int startUID, int goalUID)
 {
 	_checkingQueue.push_front(startUID);
-	GetNode(startUID).SetCost("Tentative", 0.0f);
-	while ( !_checkingQueue.empty() && _state != State::BeingCollected )
+	NodeByUid(startUID).SetCost("Tentative", 0.0f);
+	while (!_checkingQueue.empty() && _state != PathfinderState::BeingCollected)
 	{
 		PauseCheck();
 		_activeNodeUID = _checkingQueue.front();
-		if ( _activeNodeUID == goalUID )
+		if (_activeNodeUID == goalUID)
 		{
 			break;
 		}
 
-		Node &activeNode = GetNode(_activeNodeUID);
+		Node& activeNode = NodeByUid(_activeNodeUID);
 		_checkingQueue.pop_front();
 
-		for ( const auto &neighborUID : activeNode.GetNeighbors() )
+		for (const auto& neighborUID : activeNode.Neighbors())
 		{
-			if ( _state == State::BeingCollected )
-				break;
+			if (_state == PathfinderState::BeingCollected) return;
 			PauseCheck();
 			SleepDelay();
 
-			Node &neighbor = GetNode(neighborUID);
-			if ( !_traverseGrid->IsEdgeObstacle(_activeNodeUID, neighborUID) && neighborUID != activeNode.GetViaUID() )
+			Node& neighbor = NodeByUid(neighborUID);
+			if (!_traverseGrid->IsEdgeObstacle(_activeNodeUID, neighborUID) && neighborUID != activeNode.ViaUID())
 			{
-				const float suggestedTentativeCost =
-					activeNode.GetCost("Tentative") + activeNode.GetNeighborCost(neighborUID);
-				if ( suggestedTentativeCost < neighbor.GetCost("Tentative") )
+				const float suggestedTentativeCost = activeNode.Cost("Tentative") + activeNode.NeighborCostByUid(
+					neighborUID);
+				if (suggestedTentativeCost < neighbor.Cost("Tentative"))
 				{
-					if ( std::find(_checkingQueue.begin(), _checkingQueue.end(), neighborUID) == _checkingQueue.end() )
-						_checkingQueue.push_back(neighborUID);
+					if (std::ranges::find(_checkingQueue, neighborUID) == _checkingQueue.end()) _checkingQueue.
+						push_back(neighborUID);
 
 					neighbor.SetVia(_activeNodeUID);
 					neighbor.SetCost("Tentative", suggestedTentativeCost);
@@ -41,10 +46,10 @@ void Dijkstra::FindPath(int startUID, int goalUID)
 			}
 			activeNode.AddVisitedNeighbor(neighborUID);
 		}
-		std::sort(_checkingQueue.begin(), _checkingQueue.end(), [this](const auto &lhs, const auto &rhs)
-				  {
-					  return GetNode(lhs).GetCost("Tentative") < GetNode(rhs).GetCost("Tentative");
-				  });
+		std::ranges::sort(_checkingQueue, [this](const auto& lhs, const auto& rhs)
+		{
+			return NodeByUid(lhs).Cost("Tentative") < NodeByUid(rhs).Cost("Tentative");
+		});
 	}
 	_checkingQueue.clear();
 }
